+++ examples/animals.py
@@ -2,7 +2,7 @@ import sys
 
 from pydantic import BaseModel
 
-from src import MermaidGenerator, Relations
+from pydantic_mermaid import MermaidGenerator, Relations
 
 
 class Animal(BaseModel):