+++ pydantic_mermaid/pydantic_parser.py
@@ -1,0 +1,1 @@
+from pydantic_mermaid.pydantic_2_parser import Pydantic2Parser as PydanticParser  # noqa