+++ src/pydantic_parser.py
@@ -1,1 +1,0 @@
-from .pydantic_2_parser import Pydantic2Parser as PydanticParser  # noqa