+++ tests/test_contry.py
@@ -1,6 +1,6 @@
 from examples import country
 
-from src import MermaidGenerator, Relations
+from pydantic_mermaid import MermaidGenerator, Relations
 
 
 def test_country():